Move save component logic into ComponentSettings
