+++ frontend/form-builder/src/components/FormSettings/SettingsModal.tsx
@@ -2,13 +2,11 @@ import React, { useContext } from "react";
 import { Modal } from "@digdir/design-system-react";
 import classes from "./FormSettings.module.css";
 import { FormBuilderContext } from "../../context";
+import { SettingsContent } from "./SettingsContent";
 
-interface SettingsSidebarProps {
-  SettingsContent: () => React.JSX.Element;
-}
-
-export const SettingsModal = ({ SettingsContent }: SettingsSidebarProps): React.JSX.Element => {
+export const SettingsModal = (): React.JSX.Element => {
   const { modalRef } = useContext(FormBuilderContext);
+
   return (
     <Modal ref={modalRef} className={classes.modalWindow}>
       <Modal.Header className={classes.modalHeader}>Settings</Modal.Header>
